@@ -2,7 +2,7 @@
     "name": "qs",
     "description": "A querystring parser that supports nesting and arrays, with a depth limit",
     "homepage": "https://github.com/ljharb/qs",
-    "version": "6.9.3",
+    "version": "6.11.2",
     "repository": {
         "type": "git",
         "url": "https://github.com/ljharb/qs.git"
@@ -29,41 +29,52 @@
     "engines": {
         "node": ">=0.6"
     },
-    "dependencies": {},
+    "dependencies": {
+        "side-channel": "^1.0.4"
+    },
     "devDependencies": {
-        "@ljharb/eslint-config": "^16.0.0",
-        "browserify": "^16.5.0",
-        "covert": "^1.1.1",
+        "@ljharb/eslint-config": "^21.0.1",
+        "aud": "^2.0.2",
+        "browserify": "^16.5.2",
         "eclint": "^2.8.1",
-        "eslint": "^6.8.0",
+        "eslint": "=8.8.0",
         "evalmd": "^0.0.19",
         "for-each": "^0.3.3",
-        "has-symbols": "^1.0.1",
+        "has-override-mistake": "^1.0.0",
+        "has-property-descriptors": "^1.0.0",
+        "has-symbols": "^1.0.3",
         "iconv-lite": "^0.5.1",
-        "mkdirp": "^0.5.4",
-        "object-inspect": "^1.7.0",
+        "in-publish": "^2.0.1",
+        "mkdirp": "^0.5.5",
+        "mock-property": "^1.0.0",
+        "npmignore": "^0.3.0",
+        "nyc": "^10.3.2",
+        "object-inspect": "^1.12.3",
         "qs-iconv": "^1.0.4",
-        "safe-publish-latest": "^1.1.4",
+        "safe-publish-latest": "^2.0.0",
         "safer-buffer": "^2.1.2",
-        "tape": "^5.0.0-next.5"
+        "tape": "^5.6.3"
     },
     "scripts": {
-        "prepublish": "safe-publish-latest && npm run dist",
+        "prepack": "npmignore --auto --commentLines=autogenerated",
+        "prepublishOnly": "safe-publish-latest && npm run dist",
+        "prepublish": "not-in-publish || npm run prepublishOnly",
         "pretest": "npm run --silent readme && npm run --silent lint",
-        "test": "npm run --silent coverage",
-        "tests-only": "node test",
-        "posttest": "npx aud --production",
+        "test": "npm run tests-only",
+        "tests-only": "nyc tape 'test/**/*.js'",
+        "posttest": "aud --production",
         "readme": "evalmd README.md",
-        "postlint": "eclint check * lib/* test/*",
-        "lint": "eslint lib/*.js test/*.js",
-        "coverage": "covert test",
+        "postlint": "eclint check $(git ls-files | xargs find 2> /dev/null | grep -vE 'node_modules|\\.git' | grep -v dist/)",
+        "lint": "eslint --ext=js,mjs .",
         "dist": "mkdirp dist && browserify --standalone Qs lib/index.js > dist/qs.js"
     },
     "license": "BSD-3-Clause",
-    "greenkeeper": {
+    "publishConfig": {
         "ignore": [
-            "iconv-lite",
-            "mkdirp"
+            "!dist/*",
+            "bower.json",
+            "component.json",
+            ".github/workflows"
         ]
     }
 }
